chore: setup authentication

+++ backend/src/services/posts.js
@@ -1,15 +1,16 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
-// CreatePost service function =========================================
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+// CreatePost service function ============================================
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
 
   return await post.save()
 }
 
-// Delete Post service function ========================================
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+// Delete Post service function ===========================================
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
 
 // Get Post service function ===============================================
@@ -27,19 +28,24 @@ async function listPosts(
 export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+
+// List all posts by the author =============================================
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
+
 export async function listPostsByTag(tags, options) {
   return await listPosts({ tags }, options)
 }
 
-//
 // Update a post==============================================================
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(userId, postId, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
